backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pydantic
pydantic-settings
python-multipart
django
